Added image/video autofit and background selector Added automatical content type detection and corresponding content icon in index.html Removed url validation function Added url scheme guessing Corrected volume type from int to float Corrected callbacks on asset removal Added content-type==image in electron main page

+++ manager/utils/audio.py
@@ -24,7 +24,7 @@ def setDefaultAudioDevice(dev: str):
     with rpyc.connect('localhost') as conn, conn.modules.pulsectl.Pulse() as pulse:
         pulse.default_set(pulse.get_sink_by_name(dev))
 
-def setVolume(dev: str|None, volume: int):
+def setVolume(dev: str|None, volume: float):
     with rpyc.connect('localhost') as conn, conn.modules.pulsectl.Pulse() as pulse:
         pulse.volume_set_all_chans(pulse.get_sink_by_name(
             dev or pulse.server_info().default_sink_name), volume)
